Precision issue when recording values from TClientDataSet - by magnomp - Mantis Bug 162
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing@532 d681b350-854c-439c-a81d-c0ce20710203

+++ src/component/ZAbstractRODataset.pas
@@ -3340,7 +3340,7 @@ begin
             ftInteger, ftAutoInc:
               Statement.SetInt(I + 1, ParamValue.AsInteger);
             ftFloat:
-              Statement.SetFloat(I + 1, ParamValue.AsFloat);
+              Statement.SetDouble(I + 1, ParamValue.AsFloat);
             ftLargeInt:
               Statement.SetInt(I + 1, ParamValue.AsInteger);
             ftString:
